Refactor into appropriate packages.
