--- app/src/main/java/com/mochire/tech/ui/main/conditions/ConditionsViewModel.kt
@@ -1,10 +1,0 @@
-package com.mochire.tech.ui.main.conditions
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class ConditionsViewModel : ViewModel() {
-
-
-}
