+++ app/src/main/java/com/mochire/tech/ui/main/fragments/ConditionsFragment.kt
@@ -1,13 +1,13 @@
-package com.mochire.tech.ui.main.conditions
+package com.mochire.tech.ui.main.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.mochire.tech.databinding.FragmentConditionsBinding
+import com.mochire.tech.viewmodels.ConditionsViewModel
 
 class ConditionsFragment : Fragment() {
