+++ app/src/main/java/com/mochire/tech/ui/main/fragments/ProfileFragment.kt
@@ -1,14 +1,14 @@
-package com.mochire.tech.ui.main.profile
+package com.mochire.tech.ui.main.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.google.firebase.auth.FirebaseAuth
 import com.mochire.tech.databinding.FragmentProfileBinding
+import com.mochire.tech.viewmodels.ProfileViewModel
 
 class ProfileFragment : Fragment() {
