+++ app/src/main/java/com/mochire/tech/viewmodels/ConditionsViewModel.kt
@@ -1,0 +1,8 @@
+package com.mochire.tech.viewmodels
+
+import androidx.lifecycle.ViewModel
+
+class ConditionsViewModel : ViewModel() {
+
+
+}
