+++ app/src/main/res/layout/fragment_conditions.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.conditions.ConditionsFragment">
+    tools:context=".ui.main.fragments.ConditionsFragment">
 
     <TextView
         android:id="@+id/text_notifications"
